refactor: gather functions that create or update user IAM role
This is to then be able to further tidy things up, and hopefully fix issues
when users don't get updated team permissions when their email address changes.
